fix: correcting code organization
