--- src/main/java/com/hobbyFinder/hubby/controller/AuthController.java
@@ -1,23 +1,0 @@
-package com.hobbyFinder.hubby.controller;
-
-import com.hobbyFinder.hubby.exception.AuthException.Login.CredenciaisLoginException;
-import com.hobbyFinder.hubby.exception.AuthException.Registro.CredenciaisRegistroException;
-import com.hobbyFinder.hubby.models.dto.user.AuthDTO;
-import com.hobbyFinder.hubby.models.dto.user.LoginResponseDTO;
-import com.hobbyFinder.hubby.models.dto.user.RegisterDTO;
-import com.hobbyFinder.hubby.models.entities.CustomPrincipal;
-import com.hobbyFinder.hubby.services.IServices.AuthInterface;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.web.bind.annotation.*;
-
-@RestController
-@RequestMapping("auth")
-public class AuthController {
-
-    @Autowired
-    private AuthInterface authInterface;
-}
