+++ src/main/java/com/hobbyFinder/hubby/controller/routes/EvaluationRoutes.java
@@ -1,0 +1,6 @@
+package com.hobbyFinder.hubby.controller.routes;
+
+public class EvaluationRoutes {
+    public static final String POST_EVALUATION_EVENT = EventRoutes.EVENT_BASE + "/{idEvent}" + "/evaluation";
+    public static final String GET_EVALUATION_EVENT = EventRoutes.EVENT_BASE + "/{idEvent}" + "/evaluation";
+}
