+++ src/main/java/com/hobbyFinder/hubby/controller/routes/NotificationRoutes.java
@@ -1,0 +1,4 @@
+package com.hobbyFinder.hubby.controller.routes;
+
+public class NotificationRoutes {
+}
