+++ src/test/java/com/hobbyFinder/hubby/controllerTest/ParticipationTests/GetAllParticipationsEventTest.java
@@ -1,18 +1,19 @@
-package com.hobbyFinder.hubby.controllerTest.EventTests;
+package com.hobbyFinder.hubby.controllerTest.ParticipationTests;
 
 import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
+import com.hobbyFinder.hubby.controller.routes.ParticipationRoutes;
 import com.hobbyFinder.hubby.controller.routes.UserRoutes;
-import com.hobbyFinder.hubby.controllerTest.PartiticipationRequestTests.RequestConstants;
-import com.hobbyFinder.hubby.controllerTest.UserTests.UserConstants;
+import com.hobbyFinder.hubby.controllerTest.EventTests.EventSeeder;
 import com.hobbyFinder.hubby.controllerTest.UserTests.UserSeeder;
 import com.hobbyFinder.hubby.exception.CustomErrorType;
+import com.hobbyFinder.hubby.exception.ParticipationExceptions.ParticipationExceptionsMessages;
 import com.hobbyFinder.hubby.models.dto.participationRequest.ParticipationRequestEventDto;
-import com.hobbyFinder.hubby.models.entities.Participation;
+import com.hobbyFinder.hubby.models.entities.Event;
+import com.hobbyFinder.hubby.models.enums.PrivacyEnum;
 import com.hobbyFinder.hubby.repositories.EventRepository;
 import com.hobbyFinder.hubby.repositories.ParticipationRepository;
-import com.hobbyFinder.hubby.repositories.UserRepository;
 import jakarta.transaction.Transactional;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
@@ -21,27 +22,24 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.PageImpl;
 import org.springframework.http.MediaType;
 import org.springframework.test.context.ActiveProfiles;
 import org.springframework.test.web.servlet.MockMvc;
-
-import java.util.List;
-import java.util.UUID;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
+import java.util.List;
+import java.util.UUID;
+
 @SpringBootTest
 @AutoConfigureMockMvc
 @Transactional
 @ActiveProfiles("test")
-@DisplayName("Pega todas as participações de usuario")
-
-public class GetAllParticipationsUserTest {
+@DisplayName("Pega todas as participações de evento")
+public class GetAllParticipationsEventTest {
 
     @Autowired
     private MockMvc driver;
@@ -73,12 +71,18 @@ public class GetAllParticipationsUserTest {
         this.eventRepository.deleteAll();
     }
 
+    private UUID getEventId() {
+        return eventRepository.findAll().get(0).getId();
+    }
+
     @Transactional
     @Test
-    @DisplayName("Pega todas as participações de usuário")
-    void getAllParticipationsUser() throws Exception {
+    @DisplayName("Pega todas as participações de evento com sucesso.")
+    void getAllParticipationsEventTest() throws Exception {
 
-        String responseJsonString = driver.perform(get(UserRoutes.GET_ALL_USER_PARTICIPATIONS)
+        UUID idEvent = getEventId();
+
+        String responseJsonString = driver.perform(get(ParticipationRoutes.GET_ALL_EVENT_PARTICIPATIONS, idEvent)
                         .contentType(MediaType.APPLICATION_JSON)
                         .param("size", "20")
                         .param("page", "0")
@@ -93,16 +97,34 @@ public class GetAllParticipationsUserTest {
                 new TypeReference<List<ParticipationRequestEventDto>>() {
                 });
 
-        System.out.println("Lista convertida: " + content);
-
-        assertEquals(2, content.size());
-
+        assertEquals(1, content.size());
     }
 
     @Transactional
     @Test
-    @DisplayName("Quando não existe participações do usuário.")
-    void getAllParticipationsFail() throws Exception {
+    @DisplayName("Identificador do evento não existe.")
+    void getAllParticipationsEventNotFoundTest() throws Exception {
+
+        UUID idEvent = UUID.randomUUID();
+
+        String responseJsonString = driver.perform(get(ParticipationRoutes.GET_ALL_EVENT_PARTICIPATIONS, idEvent)
+                        .contentType(MediaType.APPLICATION_JSON)
+                        .param("size", "20")
+                        .param("page", "0")
+                        .header("Authorization", "Bearer " + token))
+                .andExpect(status().isNotFound())
+                .andReturn().getResponse().getContentAsString();
+
+        CustomErrorType customErrorType = objectMapper.readValue(responseJsonString, CustomErrorType.class);
+        assertEquals(customErrorType.getMessage(), "Evento não encontrado.");
+    }
+
+    @Transactional
+    @Test
+    @DisplayName("Usuário não está presente no evento.")
+    void getAllParticipationsEventUserNotInParticipation() throws Exception {
+
+        UUID idEvent = getEventId();
 
         driver.perform(post(UserRoutes.LOGOUT)
                         .contentType(MediaType.APPLICATION_JSON)
@@ -111,15 +133,20 @@ public class GetAllParticipationsUserTest {
 
         String secondToken = userSeeder.loginSegundoUser();
 
-        String responseJsonString = driver.perform(get(UserRoutes.GET_ALL_USER_PARTICIPATIONS)
+        Event event = eventRepository.findById(idEvent).get();
+        event.setPrivacy(PrivacyEnum.PRIVATE);
+        eventRepository.save(event);
+
+        String responseJsonString = driver.perform(get(ParticipationRoutes.GET_ALL_EVENT_PARTICIPATIONS, idEvent)
                         .contentType(MediaType.APPLICATION_JSON)
                         .param("size", "20")
                         .param("page", "0")
                         .header("Authorization", "Bearer " + secondToken))
-                .andExpect(status().isNotFound())
+                .andExpect(status().isForbidden())
                 .andReturn().getResponse().getContentAsString();
 
         CustomErrorType customErrorType = objectMapper.readValue(responseJsonString, CustomErrorType.class);
-        assertEquals(customErrorType.getMessage(), "A página indicada está vazia");
+        assertEquals(customErrorType.getMessage(), ParticipationExceptionsMessages.USER_REQUISITION_DENIED);
     }
+
 }
