+++ src/test/java/com/hobbyFinder/hubby/controllerTest/ParticipationTests/UserDeleteParticipationTest.java
@@ -1,17 +1,12 @@
-package com.hobbyFinder.hubby.controllerTest.EventTests;
+package com.hobbyFinder.hubby.controllerTest.ParticipationTests;
 
-import com.fasterxml.jackson.core.type.TypeReference;
-import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.hobbyFinder.hubby.controller.routes.EventRoutes;
+import com.hobbyFinder.hubby.controller.routes.ParticipationRoutes;
 import com.hobbyFinder.hubby.controller.routes.UserRoutes;
+import com.hobbyFinder.hubby.controllerTest.EventTests.EventSeeder;
 import com.hobbyFinder.hubby.controllerTest.UserTests.UserSeeder;
 import com.hobbyFinder.hubby.exception.CustomErrorType;
-import com.hobbyFinder.hubby.exception.EventException.EventExceptionsMessages;
 import com.hobbyFinder.hubby.exception.ParticipationExceptions.ParticipationExceptionsMessages;
-import com.hobbyFinder.hubby.models.dto.participationRequest.ParticipationRequestEventDto;
-import com.hobbyFinder.hubby.models.entities.Event;
-import com.hobbyFinder.hubby.models.enums.PrivacyEnum;
 import com.hobbyFinder.hubby.repositories.EventRepository;
 import com.hobbyFinder.hubby.repositories.ParticipationRepository;
 import jakarta.transaction.Transactional;
@@ -26,20 +21,19 @@ import org.springframework.http.MediaType;
 import org.springframework.test.context.ActiveProfiles;
 import org.springframework.test.web.servlet.MockMvc;
 
+import java.util.UUID;
+
 import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.delete;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
-
-import java.util.List;
-import java.util.UUID;
 
 @SpringBootTest
 @AutoConfigureMockMvc
 @Transactional
 @ActiveProfiles("test")
-@DisplayName("Pega todas as participações de evento")
-public class GetAllParticipationsEventTest {
+@DisplayName("Testes de rota: Deleção de Participação em evento.")
+public class UserDeleteParticipationTest {
 
     @Autowired
     private MockMvc driver;
@@ -77,76 +71,82 @@ public class GetAllParticipationsEventTest {
 
     @Transactional
     @Test
-    @DisplayName("Pega todas as participações de evento com sucesso.")
-    void getAllParticipationsEventTest() throws Exception {
+    @DisplayName("User deleta sua participação com sucesso.")
+    void testSelfUserDeleteParticipation() throws Exception {
 
         UUID idEvent = getEventId();
+        UUID idParticipation = eventRepository.findById(idEvent).get().getParticipations().get(0).getIdParticipation();
 
-        String responseJsonString = driver.perform(get(EventRoutes.EVENT_BASE + "/" +idEvent + "/participations")
-                        .contentType(MediaType.APPLICATION_JSON)
-                        .param("size", "20")
-                        .param("page", "0")
-                        .header("Authorization", "Bearer " + token))
-                .andExpect(status().isOk())
-                .andReturn().getResponse().getContentAsString();
+        driver.perform(delete(ParticipationRoutes.USER_DELETE_PARTICIPATION, idEvent, idParticipation)
+                .contentType(MediaType.APPLICATION_JSON)
+                .header("Authorization", "Bearer " + token))
+                .andExpect(status().isNoContent());
 
-        JsonNode rootNode = objectMapper.readTree(responseJsonString);
-        JsonNode contentNode = rootNode.get("content");
-        List<ParticipationRequestEventDto> content = objectMapper.readValue(
-                contentNode.toString(),
-                new TypeReference<List<ParticipationRequestEventDto>>() {
-                });
-
-        assertEquals(1, content.size());
     }
 
     @Transactional
     @Test
-    @DisplayName("Identificador do evento não existe.")
-    void getAllParticipationsEventNotFoundTest() throws Exception {
+    @DisplayName("Identificador de participação não foi cadastrado")
+    void testSelfUserDeleteParticipationInvalid() throws Exception {
+        UUID idEvent = getEventId();
+        UUID idParticipation = UUID.randomUUID();
 
-        UUID idEvent = UUID.randomUUID();
-
-        String responseJsonString = driver.perform(get(EventRoutes.EVENT_BASE + "/" +idEvent + "/participations")
+        String responseJsonString = driver.perform(delete(ParticipationRoutes.USER_DELETE_PARTICIPATION, idEvent, idParticipation)
                         .contentType(MediaType.APPLICATION_JSON)
-                        .param("size", "20")
-                        .param("page", "0")
                         .header("Authorization", "Bearer " + token))
                 .andExpect(status().isNotFound())
                 .andReturn().getResponse().getContentAsString();
 
         CustomErrorType customErrorType = objectMapper.readValue(responseJsonString, CustomErrorType.class);
-        assertEquals(customErrorType.getMessage(), "Evento não encontrado.");
+        assertEquals(customErrorType.getMessage(),"Participação não encontrada!");
     }
 
     @Transactional
     @Test
-    @DisplayName("Usuário não está presente no evento.")
-    void getAllParticipationsEventUserNotInParticipation() throws Exception {
+    @DisplayName("Identificador de evento não foi cadastrado")
+    void testSelfUserDeleteUnexistentEvent() throws Exception {
+
+        UUID idEvent = UUID.randomUUID();
+        UUID idParticipation = UUID.randomUUID();
+
+        String responseJsonString = driver.perform(delete(ParticipationRoutes.USER_DELETE_PARTICIPATION, idEvent, idParticipation)
+                        .contentType(MediaType.APPLICATION_JSON)
+                        .header("Authorization", "Bearer " + token))
+                .andExpect(status().isNotFound())
+                .andReturn().getResponse().getContentAsString();
+
+        CustomErrorType customErrorType = objectMapper.readValue(responseJsonString, CustomErrorType.class);
+        assertEquals(customErrorType.getMessage(),"Evento não encontrado.");
+
+    }
+
+    //Esse teste só existe por causa de uma singularidade do código, pois foi pedido para ser armazenado ids.
+    @Transactional
+    @Test
+    @DisplayName("O usuário está tentando excluir a participação de outro user.")
+    void testSelfDeleteIncorrectUser() throws Exception {
 
         UUID idEvent = getEventId();
+        UUID idParticipation = eventRepository.findById(idEvent).get().getParticipations().get(0).getIdParticipation();
 
+        //Fazendo logout do usuário que criou o evento
         driver.perform(post(UserRoutes.LOGOUT)
                         .contentType(MediaType.APPLICATION_JSON)
                         .header("Authorization", "Bearer " + token))
                 .andExpect(status().isNoContent());
 
-        String secondToken = userSeeder.loginSegundoUser();
+        //Logando o usuário incorreto
+        String incorrectToken = userSeeder.loginSegundoUser();
 
-        Event event = eventRepository.findById(idEvent).get();
-        event.setPrivacy(PrivacyEnum.PRIVATE);
-        eventRepository.save(event);
-
-        String responseJsonString = driver.perform(get(EventRoutes.EVENT_BASE + "/" +idEvent + "/participations")
+        String responseJsonString = driver.perform(delete(ParticipationRoutes.USER_DELETE_PARTICIPATION, idEvent, idParticipation)
                         .contentType(MediaType.APPLICATION_JSON)
-                        .param("size", "20")
-                        .param("page", "0")
-                        .header("Authorization", "Bearer " + secondToken))
-                .andExpect(status().isForbidden())
+                        .header("Authorization", "Bearer " + incorrectToken))
+                .andExpect(status().isConflict())
                 .andReturn().getResponse().getContentAsString();
 
         CustomErrorType customErrorType = objectMapper.readValue(responseJsonString, CustomErrorType.class);
-        assertEquals(customErrorType.getMessage(), ParticipationExceptionsMessages.USER_REQUISITION_DENIED);
+        assertEquals(customErrorType.getMessage(),ParticipationExceptionsMessages.SELF_DELETE_ID_DENIED);
+
     }
 
 }
