+++ src/test/java/com/hobbyFinder/hubby/controllerTest/ParticipationTests/UserUpdateParticipationTest.java
@@ -1,19 +1,18 @@
-package com.hobbyFinder.hubby.controllerTest.EventTests;
+package com.hobbyFinder.hubby.controllerTest.ParticipationTests;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.hobbyFinder.hubby.controller.routes.EventRoutes;
+import com.hobbyFinder.hubby.controller.routes.ParticipationRoutes;
 import com.hobbyFinder.hubby.controller.routes.UserRoutes;
+import com.hobbyFinder.hubby.controllerTest.EventTests.EventSeeder;
 import com.hobbyFinder.hubby.controllerTest.UserTests.UserConstants;
 import com.hobbyFinder.hubby.controllerTest.UserTests.UserSeeder;
 import com.hobbyFinder.hubby.exception.CustomErrorType;
-import com.hobbyFinder.hubby.exception.ParticipationExceptions.ParticipationExceptionsMessages;
 import com.hobbyFinder.hubby.models.entities.Participation;
 import com.hobbyFinder.hubby.models.enums.ParticipationPosition;
 import com.hobbyFinder.hubby.models.enums.UserParticipation;
 import com.hobbyFinder.hubby.repositories.EventRepository;
 import com.hobbyFinder.hubby.repositories.ParticipationRepository;
 import com.hobbyFinder.hubby.repositories.UserRepository;
-import com.hobbyFinder.hubby.util.GetUserLogged;
 import jakarta.transaction.Transactional;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
@@ -29,16 +28,15 @@ import org.springframework.test.web.servlet.MockMvc;
 import java.util.UUID;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.delete;
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.*;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
 @SpringBootTest
 @AutoConfigureMockMvc
 @Transactional
 @ActiveProfiles("test")
-@DisplayName("Testes de rota: Expulsão de outro usuário em evento.")
-public class DeleteUserParticipationTest {
+@DisplayName("Testes de update de participação de usuário.")
+public class UserUpdateParticipationTest {
 
     @Autowired
     private MockMvc driver;
@@ -49,12 +47,12 @@ public class DeleteUserParticipationTest {
     @Autowired
     private EventSeeder eventSeeder;
     @Autowired
+    private EventRepository eventRepository;
+    @Autowired
     private ParticipationRepository participationRepository;
     @Autowired
-    private EventRepository eventRepository;
-    private String token;
-    @Autowired
     private UserRepository userRepository;
+    private String token;
 
     @BeforeEach
     void setUp() throws Exception {
@@ -76,7 +74,7 @@ public class DeleteUserParticipationTest {
         return eventRepository.findAll().get(0).getId();
     }
 
-    private UUID createSecondUserParticipation() throws Exception {
+    private UUID createParticipation() throws Exception {
 
         UUID idEvent = getEventId();
         Participation newParticipation =
@@ -95,28 +93,41 @@ public class DeleteUserParticipationTest {
 
     @Transactional
     @Test
-    @DisplayName("Usuário expulsa outro usuário com sucesso.")
-    void testExpelSucess() throws Exception {
+    @DisplayName("Update de participação do usuário com sucesso.")
+    void updateUserParticipationSucess() throws Exception {
 
         UUID idEvent = getEventId();
-        UUID idNewParticipation = createSecondUserParticipation();
+        UUID idNewParticipation = createParticipation();
+        UserParticipation userParticipation = UserParticipation.CONFIRMED_PRESENCE;
 
-        driver.perform(delete(EventRoutes.EVENT_BASE + "/" + idEvent + "/participation/" + idNewParticipation)
-                .header("Authorization", "Bearer " + token))
+        driver.perform(post(UserRoutes.LOGOUT)
+                        .contentType(MediaType.APPLICATION_JSON)
+                        .header("Authorization", "Bearer " + token))
+                .andExpect(status().isNoContent());
+
+        String segundoUser = userSeeder.loginSegundoUser();
+
+        driver.perform(put(ParticipationRoutes.USER_UPDATE_PARTICIPATION, idEvent, idNewParticipation)
+                        .contentType(MediaType.APPLICATION_JSON)
+                        .header("Authorization", "Bearer " + segundoUser)
+                        .param("userParticipation", userParticipation.name()))
                 .andExpect(status().isNoContent());
 
     }
 
     @Transactional
     @Test
-    @DisplayName("Usuário passa um evento não cadastrado")
-    void testExpelIncorrectEvent() throws Exception {
+    @DisplayName("Identificador de evento não foi encontrado.")
+    void updateUserEventNotFound() throws Exception {
 
         UUID idEvent = UUID.randomUUID();
-        UUID idNewParticipation = createSecondUserParticipation();
+        UUID idNewParticipation = createParticipation();
+        UserParticipation userParticipation = UserParticipation.CONFIRMED_PRESENCE;
 
-        String responseJsonString = driver.perform(delete(EventRoutes.EVENT_BASE + "/" + idEvent + "/participation/" + idNewParticipation)
-                        .header("Authorization", "Bearer " + token))
+        String responseJsonString = driver.perform(put(ParticipationRoutes.USER_UPDATE_PARTICIPATION, idEvent, idNewParticipation)
+                        .contentType(MediaType.APPLICATION_JSON)
+                        .header("Authorization", "Bearer " + token)
+                        .param("userParticipation", userParticipation.name()))
                 .andExpect(status().isNotFound())
                 .andReturn().getResponse().getContentAsString();
 
@@ -127,45 +138,21 @@ public class DeleteUserParticipationTest {
 
     @Transactional
     @Test
-    @DisplayName("Identificador de participação não encontrado")
-    void testExpelIncorrectParticipation() throws Exception {
+    @DisplayName("Identificador de participação não foi encontrado.")
+    void updateUserParticipationNotFound() throws Exception {
 
         UUID idEvent = getEventId();
         UUID idNewParticipation = UUID.randomUUID();
+        UserParticipation userParticipation = UserParticipation.CONFIRMED_PRESENCE;
 
-        String responseJsonString = driver.perform(delete(EventRoutes.EVENT_BASE + "/" + idEvent + "/participation/" + idNewParticipation)
-                        .header("Authorization", "Bearer " + token))
+        String responseJsonString = driver.perform(put(ParticipationRoutes.USER_UPDATE_PARTICIPATION, idEvent, idNewParticipation)
+                        .contentType(MediaType.APPLICATION_JSON)
+                        .header("Authorization", "Bearer " + token)
+                        .param("userParticipation", userParticipation.name()))
                 .andExpect(status().isNotFound())
                 .andReturn().getResponse().getContentAsString();
 
         CustomErrorType customErrorType = objectMapper.readValue(responseJsonString, CustomErrorType.class);
         assertEquals(customErrorType.getMessage(), "Participação não encontrada!");
     }
-
-    @Transactional
-    @Test
-    @DisplayName("Cargo do usuário não é superior ao cargo de quem está deletando a participação.")
-    void testExpelIncorrectParticipationPosition() throws Exception {
-
-        UUID idEvent = getEventId();
-        UUID idNewParticipation = createSecondUserParticipation();
-
-        driver.perform(post(UserRoutes.LOGOUT)
-                        .contentType(MediaType.APPLICATION_JSON)
-                        .header("Authorization", "Bearer " + token))
-                .andExpect(status().isNoContent());
-
-        //Logando o outro usuário.
-        String segundoToken = userSeeder.loginSegundoUser();
-
-        String responseJsonString = driver.perform(delete(EventRoutes.EVENT_BASE + "/" + idEvent + "/participation/" + idNewParticipation)
-                        .header("Authorization", "Bearer " + segundoToken))
-                .andExpect(status().isUnprocessableEntity())
-                .andReturn().getResponse().getContentAsString();
-
-        CustomErrorType customErrorType = objectMapper.readValue(responseJsonString, CustomErrorType.class);
-        assertEquals(customErrorType.getMessage(), ParticipationExceptionsMessages.USER_POSITION_DENIED);
-
-    }
-
 }
