+++ src/test/java/com/hobbyFinder/hubby/controllerTest/PartiticipationRequestTests/CreateRequestTest.java
@@ -23,7 +23,7 @@ import com.hobbyFinder.hubby.exception.CustomErrorType;
 import com.hobbyFinder.hubby.exception.EventException.EventExceptionsMessages;
 import com.hobbyFinder.hubby.exception.ParticipationExceptions.ParticipationExceptionsMessages;
 import com.hobbyFinder.hubby.repositories.EventRepository;
-import com.hobbyFinder.hubby.repositories.ParticipationRequestRepository;
+import com.hobbyFinder.hubby.repositories.RequestRepository;
 
 import jakarta.transaction.Transactional;
 
@@ -53,7 +53,7 @@ public class CreateRequestTest {
     private EventRepository eventRepository;
 
     @Autowired
-    private ParticipationRequestRepository requestRepository;
+    private RequestRepository requestRepository;
 
     @BeforeEach
     void setUp() throws Exception {
