+++ src/test/java/com/hobbyFinder/hubby/controllerTest/PartiticipationRequestTests/GetAllUserRequestTest.java
@@ -23,10 +23,9 @@ import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.hobbyFinder.hubby.controllerTest.UserTests.UserSeeder;
 import com.hobbyFinder.hubby.exception.CustomErrorType;
-import com.hobbyFinder.hubby.exception.ParticipationExceptions.ParticipationExceptionsMessages;
 import com.hobbyFinder.hubby.models.dto.participationRequest.ParticipationRequestEventDto;
 import com.hobbyFinder.hubby.repositories.EventRepository;
-import com.hobbyFinder.hubby.repositories.ParticipationRequestRepository;
+import com.hobbyFinder.hubby.repositories.RequestRepository;
 
 import jakarta.transaction.Transactional;
 
@@ -56,7 +55,7 @@ public class GetAllUserRequestTest {
     private EventRepository eventRepository;
 
     @Autowired
-    private ParticipationRequestRepository requestRepository;
+    private RequestRepository requestRepository;
 
     @BeforeEach
     void setUp() throws Exception {
